fix bug where all team memberships are reset if user joins any team
If a user joins a team, the ranking positions for its memberships have to be recalculated.
Therefore TeamRankingService takes an argument to run the update only within the specific scope of given team.

Before this commit, the service did not apply that scope when resetting all ranking positions.
The second step, re-calculating and updating the ranking positions uses the scope, but all other ranking positions remain reset.

This commit adds the scope to the reset step, thus fixes the bug.