remove profile menu
Some users reported, that it was not clear enough that one can click on the profile icon on the top right.

That profile icon opens a menu that contained a link to "Profile" and one to "Sign out".

This commit moves the link to "Profile" to the regular navigation menu, and places a "Sign out" icon where the "Profile icon" used to be.